Controllers and service refactor - related packages changes
